data: move ValidationData to _legacy (in-memory-path container)
ValidationData is only ever constructed by DataManager.get_validation_data, which only the in-memory
CellFlow path calls (streaming validation uses DAGEvalAdapter). Group it with the other in-memory-only
container (TrainingData) in _legacy; _data keeps the shared containers (ConditionData/PredictionData/
ReturnData/BaseDataMixin, used by the streaming predict / condition-embedding / DataManager paths).
Re-exported from cellflow.data, so public imports are unchanged.
