+++ src/cellflow/data/__init__.py
@@ -1,6 +1,6 @@
-from cellflow.data._data import BaseDataMixin, ConditionData, PredictionData, ValidationData
+from cellflow.data._data import BaseDataMixin, ConditionData, PredictionData
 from cellflow.data._datamanager import DataManager
-from cellflow.data._legacy import PredictionSampler, TrainingData, TrainSampler, ValidationSampler
+from cellflow.data._legacy import PredictionSampler, TrainingData, TrainSampler, ValidationData, ValidationSampler
 
 __all__ = [
     "DataManager",
